+++ Media Blocks SDK/MAUI/SimplePlayer/MauiProgram.cs
@@ -1,0 +1,24 @@
+﻿using SkiaSharp.Views.Maui.Controls.Hosting;
+using VisioForge.Core.UI.MAUI;
+
+namespace Simple_Player_MB_MAUI
+{
+    public static class MauiProgram
+    {
+        public static MauiApp CreateMauiApp()
+        {
+            var builder = MauiApp.CreateBuilder();
+            builder
+                .UseMauiApp<App>()
+                .UseSkiaSharp()
+                .ConfigureMauiHandlers(handlers => handlers.AddVisioForgeHandlers())
+                .ConfigureFonts(fonts =>
+                {
+                    fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
+                    fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
+                });
+
+            return builder.Build();
+        }
+    }
+}