+++ Media Blocks SDK/MAUI/SimplePlayer/Platforms/iOS/AppDelegate.cs
@@ -1,0 +1,10 @@
+﻿using Foundation;
+
+namespace Simple_Player_MB_MAUI
+{
+    [Register("AppDelegate")]
+    public class AppDelegate : MauiUIApplicationDelegate
+    {
+        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
+    }
+}