+++ Video Capture SDK X (crossplatform)/MAUI/QRReader/MainPage.xaml.cs
@@ -157,6 +157,8 @@ namespace QRReader
                 _core?.Dispose();
                 _core = null;
             }
+
+            VisioForgeX.DestroySDK();
         }
 
         private void Core_OnError(object sender, VisioForge.Core.Types.Events.ErrorsEventArgs e)